Convert log_spec to float32 after clamping

+++ src/chatterbox/models/s3tokenizer/s3tokenizer.py
@@ -162,7 +162,7 @@ class S3Tokenizer(S3TokenizerV2):
 
         mel_spec = self._mel_filters.to(self.device, dtype=torch.float32) @ magnitudes.to(dtype=torch.float32)
 
-        log_spec = torch.clamp(mel_spec, min=1e-10).log10()
+        log_spec = torch.clamp(mel_spec, min=1e-10).log10().to(dtype=torch.float32)
         log_spec = torch.maximum(log_spec, log_spec.max() - 8.0)
         log_spec = (log_spec + 4.0) / 4.0
         return log_spec
